Split BsplineAllocator into Single/MulitBsplineAllocator
